Remove 'defaultProvider' and rename 'common' to 'pagination'
