+++ assets/src/edit-story/app/media/local/actions.js
@@ -17,20 +17,21 @@
 /**
  * Internal dependencies
  */
-import * as common from '../common/actions';
+import * as common from '../pagination/actions';
 import * as types from './types';
 
-export const fetchMediaStart = (dispatch) =>
-  common.fetchMediaStart(dispatch, 'local');
-export const fetchMediaSuccess = (dispatch) =>
-  common.fetchMediaSuccess(dispatch, 'local');
-export const fetchMediaError = (dispatch) =>
-  common.fetchMediaError(dispatch, 'local');
-export const setNextPage = (dispatch) => common.setNextPage(dispatch, 'local');
-export const updateMediaElement = (dispatch) =>
-  common.updateMediaElement(dispatch, 'local');
-export const deleteMediaElement = (dispatch) =>
-  common.deleteMediaElement(dispatch, 'local');
+export const fetchMediaStart = (dispatch) => (properties) =>
+  common.fetchMediaStart(dispatch)({ ...properties, provider: 'local' });
+export const fetchMediaSuccess = (dispatch) => (properties) =>
+  common.fetchMediaSuccess(dispatch)({ ...properties, provider: 'local' });
+export const fetchMediaError = (dispatch) => (properties) =>
+  common.fetchMediaError(dispatch)({ ...properties, provider: 'local' });
+export const setNextPage = (dispatch) => (properties) =>
+  common.setNextPage(dispatch)({ ...properties, provider: 'local' });
+export const updateMediaElement = (dispatch) => (properties) =>
+  common.updateMediaElement(dispatch)({ ...properties, provider: 'local' });
+export const deleteMediaElement = (dispatch) => (properties) =>
+  common.deleteMediaElement(dispatch)({ ...properties, provider: 'local' });
 
 export const resetFilters = (dispatch) => () => {
   dispatch({ type: types.LOCAL_MEDIA_RESET_FILTERS });
